Removed Material Textures
Removed textures from maya as this caused Uv's to disappear. Once they were removed, I then was able to change the texel density for most of the model. Large Broken Robot Frame and ASR Body materials I was unable to do, maya kept crashing when I attempted to do so.